pyproject.toml
[project]
name = "call-transcriptions"
version = "0.1.0"
description = "Transscribe and do sentiment analysis on call transcriptions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["openai-whisper>=20240930", "pydub>=0.25.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/transcriber"]
